Try to fix type sizes and stack allocation

+++ include/stddef.h
@@ -1,8 +1,8 @@
 #ifndef _STDDEF_H
 #define _STDDEF_H
 
-typedef unsigned int size_t;
-typedef int ptrdiff_t;
+typedef unsigned long long size_t;
+typedef long long ptrdiff_t;
 
 #define NULL 0
